3.4-django-testing/django_testing/requirements.txt
django~=4.0.4
djangorestframework~=3.13.1
psycopg2-binary

pytest~=7.1.2
django-filter~=21.1